♻️ refactor(sandbox): inline hostenv package into pkg/sandbox
Move host-execution helpers from pkg/sandbox/hostenv/ into
pkg/sandbox/hostenv.go, eliminating a single-purpose sub-package.
Functions renamed with HostEnv prefix to stay unambiguous in the
broader sandbox namespace.
