+++ internal/agent/sandbox/backend_test.go
@@ -10,7 +10,6 @@ import (
 	oauth "github.com/CherryHQ/stella/internal/credentials/oauth"
 	pkgplugins "github.com/CherryHQ/stella/pkg/plugins"
 	pkgsandbox "github.com/CherryHQ/stella/pkg/sandbox"
-	"github.com/CherryHQ/stella/pkg/sandbox/hostenv"
 )
 
 // stubVaultLoader is a test-only VaultEnvLoader that returns a fixed map.
@@ -98,7 +97,7 @@ func TestCopyLocalHostEnvAllowlist(t *testing.T) {
 	t.Setenv("HTTPS_PROXY", "http://proxy.example:8080")
 
 	env := map[string]string{}
-	hostenv.CopyHostEnv(env)
+	pkgsandbox.HostEnvCopy(env)
 
 	if _, ok := env["STELLA_TEST_SECRET"]; ok {
 		t.Fatal("local sandbox env copied non-allowlisted host variable")
@@ -122,12 +121,12 @@ func TestLocalSandboxPathAllowed(t *testing.T) {
 		"/nix/store/abc/bin",
 		"/run/current-system/sw/bin",
 	} {
-		if !hostenv.PathAllowed(entry, stellaBin) {
+		if !pkgsandbox.HostEnvPathAllowed(entry, stellaBin) {
 			t.Fatalf("expected %q to be allowed", entry)
 		}
 	}
 	for _, entry := range []string{"", "/home/me/bin", "/tmp/bin", "/binary"} {
-		if hostenv.PathAllowed(entry, stellaBin) {
+		if pkgsandbox.HostEnvPathAllowed(entry, stellaBin) {
 			t.Fatalf("expected %q to be rejected", entry)
 		}
 	}
